chung

+++ mxh/src/main/java/com/baconbao/mxh/Repository/UserRepository.java
@@ -13,8 +13,7 @@ import jakarta.transaction.Transactional;
 public interface UserRepository extends JpaRepository<User, Long> {
   User findByEmail(String email);
 
-  // @Modifying
-  // @Transactional
+  
   @Query(value = "SELECT MAX(id_user) FROM user", nativeQuery = true)
-  Long countById();
+  Integer countById();
 }
